Generate the Axis stubs using the axistools-maven-plugin
The mappings are only activated through a profile and are verified to
have the exact same behaviour as the previous one. Contrary to Maven
conventions, we are still generating into the source tree, as I don't
have the time right now to move this.

The lib directory and the build.xml file are now gone, and the core
plugin jar size has dropped from 571k to 198k.

Fixes #161: Move WSDL generation into maven build