Read the repository due date from the task attributes
Instead of reading the repository configuration, potentially accessing
the network, we simply read the due date from the task, if set.

This also correct the check, since even if the repository due date
support was enabled, the task might not have a set due date.

Fixes #172: Activating a task blocks the main thread if the due date needs to be read and the configuration is stale