Store the subtasks enablement flag in the task repository
The task is now stored in the repository, and this means that we
no longer depend on the task being fully initialized. But on the
other hand we depend on internals.

The MantisCacheData structure version has been updated, to make sure
that the data is refreshed and the repository configuration updated.

As a side effect, the 'download sub-tasks' configuration property is 
removed. It never had any effect.

Fixes #167: Need to manually sync task to enable creation of sub tasks