Return null from MantisTaskDataHandler if ticket is not found
Altough not documented, it seems that the expectation is that when
a task can not be found by id, the repository connector should return
null rather than throw an exception. This seems to play nice with
the search dialog and also has a reaonsable result for an open
task editor.

181: Search for a nonexistent ticket produces a scary error dialog